Moved logging to it's own package. Fixed the rest of the comments I could find.
