+++ core/src/main/java/dk/alexandra/fresco/logging/BatchEvaluationLoggingDecorator.java
@@ -1,10 +1,10 @@
-package dk.alexandra.fresco.framework.sce.evaluator;
+package dk.alexandra.fresco.logging;
 
-import dk.alexandra.fresco.framework.PerformanceLogger;
 import dk.alexandra.fresco.framework.ProtocolCollection;
 import dk.alexandra.fresco.framework.builder.ProtocolBuilder;
 import dk.alexandra.fresco.framework.network.SCENetwork;
 import dk.alexandra.fresco.framework.network.SCENetworkSupplier;
+import dk.alexandra.fresco.framework.sce.evaluator.BatchEvaluationStrategy;
 import dk.alexandra.fresco.framework.sce.resources.ResourcePool;
 import java.io.IOException;
